Crate proc_macro_tokens

🔬 This is a nightly-only experimental API. (rustc_private)

Proc_Macro

A library for procedural macro writers.

Usage

This crate provides the prelude (at libproc_macro_tokens::prelude), which provides a number of operations: - concat, for concatenating two TokenStreams. - ident_eq, for checking if two identifiers are equal regardless of syntax context. - str_to_token_ident, for converting an &str into a Token. - keyword_to_token_delim, for converting a parse::token::keywords::Keyword into a Token. - build_delimited, for creating a new TokenStream from an existing one and a delimiter by wrapping the TokenStream in the delimiter. - build_bracket_delimited, build_brace_delimited, and build_paren_delimited, for easing the above. - build_empty_args, which returns a TokenStream containing (). - lex, which takes an &str and returns the TokenStream it represents.

TokenStreams

TokenStreams serve as the basis of the macro system. They are, in essence, vectors of TokenTrees, where indexing treats delimited values as a single term. That is, the term even(a+c) && even(b) will be indexibly encoded as even | (a+c) | even | (b) where, in reality, (a+c) is actually a decorated pointer to a | + | c.

If a user has a TokenStream that is a single, delimited value, they can use maybe_delimited to destruct it and receive the internal vector as a new TokenStream as: `(a+c)`.maybe_delimited() ~> Some(a | + | c)`

Check the TokenStream documentation for more information; the structure also provides cheap concatenation and slicing.

Modules

build 	 [
🔬 This is a nightly-only experimental API.
]

parse 	 [
🔬 This is a nightly-only experimental API.
]
Parsing utilities for writing procedural macros.

prelude 	 [
🔬 This is a nightly-only experimental API.
]